+++ vite.config.ts
@@ -1,14 +1,8 @@
 import { defineConfig } from 'vite';
 import { resolve } from 'path';
-import dts from 'vite-plugin-dts';
 
 export default defineConfig({
-  plugins: [
-    dts({
-      insertTypesEntry: true,
-      rollupTypes: true, // Consolida todos os tipos em um único arquivo .d.ts
-    }),
-  ],
+  plugins: [],
   build: {
     lib: {
       entry: resolve(__dirname, 'src/index.ts'),